[Refactor] Jwt 인증 인가 로직 리팩토링

+++ src/main/java/com/soma/snackexercise/util/RedisUtil.java
@@ -1,7 +1,6 @@
 package com.soma.snackexercise.util;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
 import org.springframework.stereotype.Component;
@@ -14,7 +13,7 @@ public class RedisUtil {
     private final RedisTemplate<String, Object> redisTemplate;
     private final RedisTemplate<String, Object> redisBlackListTemplate;
 
-    public void set(String key, Object o, Long minutes) {
+    public void set(String key, Object o, Integer minutes) {
         redisTemplate.setValueSerializer(new Jackson2JsonRedisSerializer(o.getClass()));
         redisTemplate.opsForValue().set(key, o, minutes, TimeUnit.MINUTES);
     }
@@ -31,7 +30,7 @@ public class RedisUtil {
         return redisTemplate.hasKey(key);
     }
 
-    public void setBlackList(String key, Object o, Long milliSeconds) {
+    public void setBlackList(String key, Object o, Integer milliSeconds) {
         redisBlackListTemplate.setValueSerializer(new Jackson2JsonRedisSerializer(o.getClass()));
         redisBlackListTemplate.opsForValue().set(key, o, milliSeconds, TimeUnit.MILLISECONDS);
     }
